[DDH-32] Add LayoutTemplate component
